Fix syntax highlighting in code blocks
While other forges may support `sh-session` for syntax highlighting
GitHub is not one of those forges. To get the expected highlighting you
need a) `console` as the syntax and b) some sigil to start the line to
indicated that it’s console _input_ and not _output_. This commit
normalizes these code blocks. Additionally, seeing the problems folks on
`zsh` seem to have with the `#`, it’s better to string escape the that
part of the shell so it’s compatible with `bash`, `dash`, `fish`, and
`zsh`.

Also, one Nix block was missing a `nix` highlight.